requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
# NOTE: torch must be installed separately on Jetson — see setup.sh
# Do NOT add 'torch' here; the PyPI version is CPU-only on aarch64.
transformers>=4.40.0
accelerate>=0.28.0
Pillow>=10.0.0
requests>=2.31.0
pydantic>=2.5.0
huggingface-hub>=0.21.0
sentencepiece
protobuf
